removed unnecessary imports; refactored code

+++ src/model/FreeRoom.java
@@ -3,6 +3,6 @@ package model;
 public class FreeRoom extends Room {
 
     public FreeRoom(String roomNumber, RoomType roomType) {
-        super(roomNumber, 0.00, roomType);
+        super(roomNumber, (double) 0, roomType);
     }
 }
